backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
requests
google-generativeai
python-dotenv
